+++ configs/react.js
@@ -1,10 +1,18 @@
+const projectPkg = require(process.cwd() + '/package.json');
+let reactVersion =
+  (projectPkg.dependencies || {}).react || (projectPkg.devDependencies || {}).react;
+if (reactVersion) {
+  const m = reactVersion.match(/(\d+)/);
+  reactVersion = m && m[1];
+}
+
 module.exports = {
   extends: ['plugin:react/recommended'],
   settings: {
     react: {
       // createClass: 'createClass', // Regex for Component Factory to use, default to 'createClass'
       // pragma: 'React',  // Pragma to use, default to 'React'
-      version: '15', // React version, default to the latest React stable release
+      version: reactVersion || undefined, // React version, default to the latest React stable release
     },
   },
   plugins: ['react'],